refactor(terminal): extract chrome CALayers into TerminalChromeLayers
Moves the 7 decorative CALayers (noise, content backdrop / strokes / gloss,
sidebar glow + bridge) out of TerminalContainerView into a self-contained
helper class. The host now passes a FrameInputs struct describing layout
state instead of letting the chrome read host properties directly.

Behavior preserved: same z-order, same theme/frame logic, same animation
wrapping. Net reduction: ~150 lines from TerminalContainerView.
